requirements.txt
pyyaml
transformers
peft
adapters
deepspeed
deepspeed-chat
awq
llm-awq
auto-gptq
squeezellm
pippy